You've got a friend in me...
Next time, I need to finish coding the friend. I'm planning on having him sit on top of the player's head after they meet.

+++ Assets/scripts/store/store.cs
@@ -9,7 +9,7 @@ public class store : MonoBehaviour
     public GameObject[] store_items;
     public bool can_buy_item;
     public bool cbi;
-    public int feather_count_st;
+    public float feather_count_st;
     public string[] names_of_items = {
         "water",
         "hook",
